media: i2c: ov01a10: Add ov01a1s support
The ov01a1s sensor is the exact same sensor as the ov01a10 with a different
(RGBI) color-filter.

Signed-off-by: Hans de Goede <hansg@kernel.org>
---
Note I'm not sure if we should apply this one as is and fixup
the reported mbus-format + add a bayer-order v4l2-control later
or if we delay merging until we can do it correct right away.