power: supply: axp288_charger: Add charge-inhibit support
Add charge-inhibit support by adding a rw status attribute and
inhibiting charging when "Not charging" or "Discharging"
is written to it.

The userspace API with status being writable is a bit weird,
but this is already done this way in the following psy drivers:
axp20x_battery.c, bq24735-charger.c, bq25980_charger.c, ip5xxx_power.c,
rt9467-charger.c, rt9471.c.

Note on systems with an AXP288 some (about 400mA depending on load)
current will be drawn from the battery when charging is disabled
even if there is an external power-supply which can provide all
the necessary current. So unfortunately one cannot simply disable
charging to keep the battery in good health when using a tablet
with an AXP288 permanently connected to an external power-supply.

Signed-off-by: Hans de Goede <hansg@kernel.org>