media: i2c: ov01a10: Register tweaks for ov01a1s model
The out of tree drivers from:
https://github.com/intel/ipu6-drivers/tree/master/drivers/media/i2c

Have a separate driver for the ov01a1s. This driver is 95% the same as
the ov01a10 driver (same sensor, different color-filters) but there are
a couple of registers for which the out of tree ov01a1s driver uses
different values.

Add a per model reg_sequence to struct ov01a10_sensor_cfg and apply
the register tweaks from the out of tree driver to the ov01a1s model.

Signed-off-by: Hans de Goede <hansg@kernel.org>